fix(membro4): corrige bugs criticos e completa sistema de batalha GUI
- Inimigo.java: declara campos spriteKey, boss e recompensa (faltavam, causavam NPE)
- FilaDeTurnos.java: remove metodo adicionar() duplicado, corrige getFila() (List->ArrayList), remove referencia inexistente a indiceTurnoAtual
- GerenciadorDeBatalha.java: reescrito para modo GUI (sem Scanner); adiciona getPersonagemAtivo(), executarTurnoAliado(), getLogCombate(), getAliados(), getInimigos(), getAliadosVivos(), getInimigosVivos(), isBatalhaEncerrada(), isVitoria()
- GameManager.java: adiciona case VITORIA no switch mudarEstado(); adiciona todos os 10 Chapeus de Palha em iniciarJogo(); corrige assinatura iniciarCombate para List<Aliado>; adiciona verificarVitoria()
- Personagem.java: getVidaAtual() e getVidaMaxima() agora retornam int (evita '150.0' na UI)
- TelaBatalha.java: processarFimBatalha() chama verificarVitoria() para ir para VITORIA; corrige escolherAlvo() com novos tipos
- TelaVitoria.java: cria tela epica de conclusao do Grand Line (novo arquivo)

+++ core/src/main/java/entidades/Personagem.java
@@ -53,13 +53,11 @@ public abstract class Personagem {
 
         return nome; 
     }
-    public float getVidaAtual() { 
-
-        return vidaAtual; 
+    public int getVidaAtual() {
+        return (int) vidaAtual;
     }
-    public float getVidaMaxima() { 
-        
-        return vidaMaxima; 
+    public int getVidaMaxima() {
+        return vidaMaxima;
     }
     public float getDefesa() {
